Add fertilizer catalog and dilution checks
