+++ data/fertilizers/fertilizer_dilution_limits.json
@@ -1,0 +1,5 @@
+{
+  "foxfarm_grow_big": 150,
+  "magriculture": 200,
+  "intrepid_granular_potash_0_0_60": 100
+}
